cpuilde - add SAP Note for VMware (#438)
2161991 - VMware vSphere configuration guidelines

Hypervisor doesn't pass through any control over P-States or C-States to the guest, the ESXi host will have full control over the Power Management of the server.
This means that the CPUIdle driver will show "none", which is correct. Even when it is showing any other option there, it will not have any control over the Power Management Stats of the physical server.